MyDRFProject/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
